==> Sintax/00_funciones_ayuda.R <==
##############################################
##     Encuesta Continua de Empleo [CR]     ##
## Creado: 27-sep-2023
## Modif : 28-sep-2023
# Objetivo: Funciones de ayuda

# Función codificación y selección de variables
ece_codif <- function(data){
  data <- data |> mutate(
    # Edad
    Edad = as.numeric(Edad),
    # Condición de actividad: 
    Condicion_actividad=recode(as.character(Condicion_actividad),
                               `1`= "1.Ocupado", 
                               `2`= "2.Desempleado", 
                               `3`= "3.Fuera de la PEA", 
                               .missing = NA_character_),
    # Sub empleo - Informalidad
    Subempleado= recode(as.character(Subempleado),
                        `0`= "1.No", 
                        `1`= "2.Sí", 
                        `9`= "3.No especificado", 
                        .missing = NA_character_),
    # Rama de actividad
    Rama_actividad= recode(as.character(Rama_actividad),
                         `1`= "Agricultura, ganadería, silvicultura y pesca", 
                         `2`= "Explotación de minas y canteras", 
                         `3`= "Industrias manufactureras", 
                         `4`= "Suministro de electricidad", 
                         `5`= "Suministro de agua, evacuación de aguas", 
                         `6`= "Construcción", 
                         `7`= "Comercio al por mayor y al por menor", 
                         `8`= "Transporte y almacenamiento", 
                         `9`= "Actividades de alojamiento", 
                         `10`= "Información y comunicaciones", 
                         `11`= "Act. Financieras y de seguros", 
                         `12`= "Act. Inmobiliarias", 
                         `13`= "Act. Profesionales, científicas y técnicas", 
                         `14`= "Act. Servicios administrativos y de apoyo", 
                         `15`= "Administración pública y defensa", 
                         `16`= "Enseñanza", 
                         `17`= "Act. Atención de la salud humana y de asistencia social", 
                         `18`= "Act. Artísticas, de entretenimiento y recreativas", 
                         `19`= "Otras actividades de servicios", 
                         `20`= "Act. de los hogares como empleadores",
                         `21`= "Act. de organizaciones y órganos extraterritoriales",
                         `99`= "99.No especificado",
                         .missing = NA_character_),
    # Educación 
    Educacion_asiste= recode(as.character(Educacion_asiste),
                         `1`= "1.Escuela",
                         `2`= "2.colegio",
                         `3`= "3.Parauniversitaria o universitaria",
                         `4`= "4.Enseñanza especial",
                         `5`= "5.Educación abierta en institutos para presentar exámenes ante el MEP",
                         `6`= "6.Otro tipo de formación no formal",
                         `7`= "7.No asiste",
                         .missing = NA_character_),
    # Motivo de no buscar empleo
    Motivo_nobusco=recode(as.character(Motivo_nobusco),
                          `1` = "1.Ya encontró trabajo", 
                          `2` = "2.Espera reinicio de operaciones o reapertura del negocio", 
                          `3` = "3.Espera respuesta de gestiones anteriores", 
                          `4` = "4.No tiene dinero para buscar", 
                          `5` = "5.Se cansó de buscar", 
                          `6` = "6.No le dan trabajo por edad, sexo, raza, discapacidad, otro", 
                          `7` = "7.No hay trabajo en la zona", 
                          `8` = "8.Sabe que en esta época del año no hay trabajo", 
                          `9` = "9.Solo trabaja cuando se lo piden", 
                          `10`= "10.Enfermedad o accidente", 
                          `11`= "11.Asiste a centro de enseñanza", 
                          `12`= "12.Tiene obligaciones familiares", 
                          `13`= "13.Tiene obligaciones personales", 
                          `14`= "14.Otras razones",
                          `99`= "14.Ignorado",
                          .missing = NA_character_),
    Seguro=recode(as.character(Seguro),
                          `1`="1. Sí",
                          `2`="2. No",
                          .missing = NA_character_),
    Sexo=recode(as.character(Sexo),
                          `1`="1. Hombre",
                          `2`="2. Mujer",
                          .missing = NA_character_),
    Estado_conyugal=recode(as.character(Estado_conyugal),
                           `1` = "1.Unión libre", 
                           `2` = "2.Casado(a)", 
                           `3` = "3.Divorciado(a)", 
                           `4` = "4.Separado(a)", 
                           `5` = "5.Viudo(a)", 
                           `6` = "6.Soltero(a)", 
                           `7` = "7.Casado(a) con cónyuge no residente", 
                           `8` = "8.Unión libre con compañero(a) del mismo sexo", 
                           `9` = "9.Ignorado", 
                           .missing = NA_character_),
    Joven_nini=recode(as.character(Joven_nini),
                           `1` = "Sí", 
                      .missing = NA_character_),
    Relacion_parentesco=recode(as.character(Relacion_parentesco),
                               `1`= "Jefe o jefa", 
                               `2`= "Conyuge", 
                               `3`= "Hijo/a", 
                               `4`= "Yerno o nuera", 
                               `5`= "Nieto/a", 
                               `6`= "Padre o madre", 
                               `7`= "Hermano o hermana", 
                               `8`= "Cuñado/a", 
                               `9`= "Otro familiar", 
                               `10`= "Otro no familiar", 
                               `11`= "Servicio sin otro lugar donde vivir o su familiar", 
                               `12`= "Pensionista sin otro lugar donde vivir o su familiar", 
                               `13`= "Nuevo jefe", 
                               `14`= "Hijastro o hijastra", 
                               `15`= "Suegro o suegra", 
                               `16`= "Unión libre con persona del mismo sexo", 
                               `17`= "Nuevo cónyuge", 
                               `18`= "Nuevo jefe(a) o cónyuge",
                               .missing = NA_character_),
    Educacion_nivel_grado= str_sub(as.character(Educacion_nivel_grado),1,1),
    Educacion_nivel_grado=recode(as.character(Educacion_nivel_grado),
                                `0`= "Ninguno",
                                `1`= "Primaria",
                                `2`= "Secundaria",
                                `3`= "Secundaria",
                                `4`= "Parauniversitario",
                                `5`= "Universidad",
                                `6`= "Especialidad",
                                `7`= "Especialidad",
                                `8`= "Maestría o Doctorado",
                                `9`= "Desconocido",
                                .missing = NA_character_),
    Desea_trabajar=recode(as.character(Desea_trabajar),
                                `1`= "1.No desea trabajar",
                                `2`= "2.Persona muy anciana",
                                `3`= "3.Por discapacidad permanente",
                                `4`= "4.Por enfermedad",
                                `5`= "5.Por obligaciones familiares",
                                `6`= "6.Por razones personales",
                                `7`= "7.Podría, pero en otro momento",
                                `8`= "8.Si podría, sin ninguna restricción",
                                `9`= "9.Si podría, con ciertas condiciones",
                               `99`= "Ignorado",
                                .missing = NA_character_)
  )
}

# Diccionario de variables ECE
# http://sistemas.inec.cr/pad5/index.php/catalog/306/data-dictionary/F7?file_name=III%20Trimestre%202023
